packages/gve-google/pyproject.toml
[tool.poetry]
name = "gve-google"
version = "0.1.0"
description = ""
authors = ["Matt <matnorri@cisco.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.0"
google-auth = "^2.0.1"
google-cloud-secret-manager = "^2.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.7b0"
flake8 = "^3.9.2"
flake8-colors = "^0.1.9"

[tool.poetry.scripts]
my-script = "gve_google.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
